Bean validation

+++ desafio-crud/src/main/java/com/devsuperior/desafio_crud/services/ClientService.java
@@ -2,10 +2,14 @@ package com.devsuperior.desafio_crud.services;
 import com.devsuperior.desafio_crud.dto.ClientDTO;
 import com.devsuperior.desafio_crud.entities.Client;
 import com.devsuperior.desafio_crud.repositories.ClientRepository;
+import com.devsuperior.desafio_crud.services.exceptions.NoSuchElementException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -18,7 +22,7 @@ public class ClientService {
     //findById method
     @Transactional(readOnly = true)
     public ClientDTO findById (Long id) {
-        Client result = repository.findById(id).get();
+        Client result = repository.findById(id).orElseThrow(()-> new NoSuchElementException("Id não encontrado"));
         ClientDTO dto = new ClientDTO(result);
         return dto;
     }
@@ -33,24 +37,32 @@ public class ClientService {
     //insert method
     @Transactional
     public ClientDTO insert (ClientDTO dto) {
-        Client entity = new Client();
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+            Client entity = new Client();
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
     }
 
     //update method
     @Transactional
     public ClientDTO update (Long id, ClientDTO dto) {
-        Client entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+        }
+        catch (EntityNotFoundException e) {
+             throw new NoSuchElementException("Id não encontrado");
+        }
     }
 
     //delete method
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete (Long id) {
+        if(!repository.existsById(id)) {
+            throw new NoSuchElementException("Id não encontrado");
+        }
         repository.deleteById(id);
     }
